Point: Use doubles for coordinates.
* Changed Point's members x and y to double.
* Tried to remove every (Long) cast, nextInt(), ==, ...
* Most likely missed some longs, so don't expect everything
  to work fine right now. Most notably selectRandomTriangleBySize()
  uses longs to represent a Triangle's size, which I don't understand
  anyway. Still, it is now possible to generate a polygon with
  500+ points (eg with Space Partitioning) _and_ still require it to
  be in GP. Pretty nifty.

+++ project/src/polygonsSWP/geometry/OrderedListPolygon.java
@@ -406,7 +406,7 @@ public class OrderedListPolygon
         double x = u.v1 * randomPoint1 + v.v1 * randomPoint2;
         double y = u.v2 * randomPoint1 + v.v2 * randomPoint2;
 
-        retval = new Point((long) x, (long) y);
+        retval = new Point(x, y);
 
       }
       while (!containsPoint(retval, true));
